cleaned up comment, removed unused libs

+++ src/predict.py
@@ -5,7 +5,6 @@ from tqdm.notebook import tqdm
 from dataset import ETCIDataset
 from model import create_single_unet
 import config
-from utils import get_etci_df, cleanup_etci_data
 
 
 def predict(test_df):
@@ -18,7 +17,6 @@ def predict(test_df):
     model = create_single_unet()
     model.load_state_dict(torch.load(f"{config.output_dir}/single_unet.pt"))
     model.to(device)
-    model.double()
     model.eval()
     try:
         with torch.no_grad():
@@ -31,6 +29,5 @@ def predict(test_df):
     except Exception as te:
         print(f"An exception occurred during inference: {te}")
 
-    # produce a single array of prediction from all batches of test data
-    final_predictions = np.concatenate(final_predictions, axis=0)
+    final_predictions = np.concatenate(final_predictions, axis=0)  # a single array of prediction from all batches
     return final_predictions
